move dateContext utility functions into DateUtils.js
